Update training scripts to use TrainingJob stuff
